Update to Python 3.6
The API can now be based around an async generator that waits for
several events and yields them one by one. No more need of the

async for events in monitor:
    for event in events:
	...
nonsense.

Also os.fspath removes a lot of boilerplate when working with paths.

And f-strings.

Also separated the inotify parsing bits and added a few more types of
exceptions and simplified the tests somewhat.